+++ packages/ui/src/theme/tokens/components/divider.ts
@@ -1,4 +1,37 @@
-export const divider = {
+import {
+  BackgroundColorValue,
+  BorderColorValue,
+  BorderStyleValue,
+  BorderWidthValue,
+  ColorValue,
+  DesignToken,
+  FontSizeValue,
+  OpacityValue,
+  SpaceValue,
+} from '../types/designToken';
+
+interface DividerSizeTokens {
+  borderWidth: DesignToken<BorderWidthValue>;
+}
+
+interface DividerLabelTokens {
+  color: DesignToken<ColorValue>;
+  paddingInline: DesignToken<SpaceValue>;
+  fontSize: DesignToken<FontSizeValue>;
+  backgroundColor: DesignToken<BackgroundColorValue>;
+}
+
+export interface DividerTokens {
+  borderStyle: DesignToken<BorderStyleValue>;
+  borderColor: DesignToken<BorderColorValue>;
+  borderWidth: DesignToken<BorderWidthValue>;
+  label: DividerLabelTokens;
+  small: DividerSizeTokens;
+  large: DividerSizeTokens;
+  opacity: DesignToken<OpacityValue>;
+}
+
+export const divider: DividerTokens = {
   borderStyle: { value: 'solid' },
   borderColor: { value: '{colors.border.primary.value}' },
   borderWidth: { value: '{borderWidths.medium.value}' },